Added network vibration
Asmdef finished

+++ Assets/_Main/Scripts/ScreenFlow/GameStats/StatsTools.cs
@@ -8,7 +8,7 @@
         public uint AbilityUseAmount;
         public uint GamesPlayed;
         public uint DeflectStreak;
-        public float LongestTime;
+        public uint LongestTime;
         public uint HighScore;
         public uint TotalScore;
     }
